Closes #663 Smart compression in the plugin (#664)
* remove the exif option from the settings page and its usage

* Update labels for levels

* use 2 as default value

* Update labels for levels

* show one link instead of multiple depending on compression level

* update default optimization level to 2

* update optimization level to 2 when upgrading to 2.0

* deprecate opcache reset functions

* use correct action for re-optimize link

* update version number to 2.0

* remove compression selector on bulk optimization
