+++ views/part-bulk-optimization-table-row-folder-type.php
@@ -1,7 +1,5 @@
 <?php
 defined( 'ABSPATH' ) || die( 'Cheatin’ uh?' );
-
-$default_level = Imagify_Options::get_instance()->get( 'optimization_level' );
 ?>
 <tr class="imagify-row-folder-type" data-group-id="<?php echo $data['group_id']; ?>" data-context="<?php echo $data['context']; ?>">
 	<td class="imagify-cell-checkbox">
@@ -26,20 +24,6 @@ $default_level = Imagify_Options::get_instance()->get( 'optimization_level' );
 	</td>
 	<td class="imagify-cell-original-size-size">
 		<?php echo $data['original-size']; ?>
-	</td>
-	<td class="imagify-cell-level">
-		<?php
-		$this->print_template( 'input/selector', [
-			'current_label' => __( 'Current level:', 'imagify' ),
-			'name'          => 'level[' . $data['group_id'] . ']',
-			'value'         => $default_level,
-			'values'        => [
-				0 => imagify_get_optimization_level_label( 0, '%ICON% %s' ),
-				1 => imagify_get_optimization_level_label( 1, '%ICON% %s' ),
-				2 => imagify_get_optimization_level_label( 2, '%ICON% %s' ),
-			],
-		] );
-		?>
 	</td>
 </tr>
