+++ views/part-bulk-optimization-table.php
@@ -45,7 +45,6 @@ defined( 'ABSPATH' ) || die( 'Cheatin’ uh?' );
 						<th class="imagify-cell-count-errors"><?php esc_html_e( 'Errors', 'imagify' ); ?></th>
 						<th class="imagify-cell-optimized-size-size"><?php esc_html_e( 'Optimized Size', 'imagify' ); ?></th>
 						<th class="imagify-cell-original-size-size"><?php esc_html_e( 'Original Size', 'imagify' ); ?></th>
-						<th class="imagify-cell-level"><?php esc_html_e( 'Level Selection', 'imagify' ); ?></th>
 					</tr>
 				</thead>
 				<tbody>
